+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Changed `listSeparatorColor` default from `Colors.transparent` to `BrightnessColor.bwa(alpha: 0.08)` and changed `SearchTextField` icon colors to be dependant on the brightness.
 * Added documentation for the `BrightnessColor` and `ThemedColor` models.
 * Added `contextualize` method for `Color` class.
+* Added `ContextualColor` model and its documentation.
 
 ## 0.0.1